# Simplify algorithm performance update