--- a/lib/Makefile
+++ b/lib/Makefile
@@ -63,8 +63,8 @@ coverity:
 
 # Examples using just the tables API.
 table_example: table_example.c
-	h5cc -o table_example table_example.c tables.c object_heap.c avl.c util.c
+	h5cc -o table_example table_example.c tables.c util.c
 
 # Make sure that all the files are compiled as C code.
-table_example_cpp: table_example_cpp.cc tables.o object_heap.o util.o
-	h5c++ -std=c++11 -o table_example_cpp table_example_cpp.cc tables.o object_heap.o avl.o util.o
+table_example_cpp: table_example_cpp.cc tables.o util.o
+	h5c++ -std=c++11 -o table_example_cpp table_example_cpp.cc tables.o util.o