Updates the simplify implementation to use recent algorithm improvements. 
cc @petrelharp and @molpopgen --- this is a WIP; I'll update here when there's something worth taking a look at.