@petrelharp 2018-04-10
Nice; looking forward to it.