@jeromekelleher 2018-04-11
OK! This is ready for action I think. Would you mind taking a look please @petrelharp?

There's two basic differences between this and the version we've written up in the paper:

1) I've decoupled the algorithms for doing the actual simplification from the algorithm for finding overlapping segments. The overlapping segments algorithm should be quite efficient now --- something like O(n log n) (for sorting n segments) + O(n * average number of overlapping segments). We could improve the second term a bit probably, but I think the expected number of overlapping segments is going to be quite small usually.

2) I've greatly simplified the mutation mapping code. It turns out that you can use the saved ancestry mappings to do the job of mapping the mutations to their output nodes for you. This means we can leave all the mutation mapping stuff until the end, which is WAY faster.

We get a pretty decent performance increase here. For a fairly big messy tree sequence I have lying around we get

**before:**
no sites: 85s
sites: 277s

**after**
no sites: 23s
sites: 23s

Now, there's no measurable difference between having sites or not. I'll have another quick look to see if there's any low-hanging fruit now, but I think we can start drawing a line under simplify soon.

Also, I've updated simplify to check for sites with duplicate positions and throw an error now.