@jeromekelleher 2018-04-11
Update: it looks like the majority of the time is being spent in sorting for the ``squash_edges`` function. We can probably do better here but I'm not sure it's worth the bother. Quite a lot of the 20s above is actually overhead from creating a new tree sequence.

@molpopgen, you did quite a bit of perf work with your version of simplify. Would you mind taking this for a spin and see how it performs on your workloads?