@molpopgen 2018-04-11
This sounds great.  I'll take a look soon, too.     I've been planning on
getting back to my C++ implementation, where I had also realized that doing
mutations after the fact was going to be both faster and a lot less
mentally taxing.

On Wed, Apr 11, 2018 at 9:52 AM Jerome Kelleher <notifications@github.com>
wrote:

> OK! This is ready for action I think. Would you mind taking a look please
> @petrelharp <https://github.com/petrelharp>?
>
> There's two basic differences between this and the version we've written
> up in the paper:
>
>    1.
>
>    I've decoupled the algorithms for doing the actual simplification from
>    the algorithm for finding overlapping segments. The overlapping segments
>    algorithm should be quite efficient now --- something like O(n log n) (for
>    sorting n segments) + O(n * average number of overlapping segments). We
>    could improve the second term a bit probably, but I think the expected
>    number of overlapping segments is going to be quite small usually.
>    2.
>
>    I've greatly simplified the mutation mapping code. It turns out that
>    you can use the saved ancestry mappings to do the job of mapping the
>    mutations to their output nodes for you. This means we can leave all the
>    mutation mapping stuff until the end, which is WAY faster.
>
> We get a pretty decent performance increase here. For a fairly big messy
> tree sequence I have lying around we get
>
> *before:*
> no sites: 85s
> sites: 277s
>
> *after*
> no sites: 23s
> sites: 23s
>
> Now, there's no measurable difference between having sites or not. I'll
> have another quick look to see if there's any low-hanging fruit now, but I
> think we can start drawing a line under simplify soon.
>
> Also, I've updated simplify to check for sites with duplicate positions
> and throw an error now.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/tskit-dev/msprime/pull/453#issuecomment-380518529>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AGHnH25VlkT0_ySemXx__xG9Qhc715zTks5tnjJbgaJpZM4TOIWB>
> .
>
-- 

Kevin Thornton

Associate Professor

Ecology and Evolutionary Biology

UC Irvine

http://www.molpopgen.org

http://github.com/ThorntonLab

http://github.com/molpopgen