@molpopgen 2018-04-12
This sounds good.  I'll ask David to run some tests with this branch.

On Thu, Apr 12, 2018 at 4:24 AM Jerome Kelleher <notifications@github.com>
wrote:

> I thought of a way to speed up the edge squashing bit and implemented it.
> It makes things faster. It's at the point where pretty much everything I
> throw at it is taking < 3 seconds now (on a reasonably fast computer), so I
> think we're pretty good. I'm interested to see how it performs on the
> forward sims use case now though. The improvements should be significant.
>
> Also worth noting here is that "tables.c" no longer depends on avl.c,
> which is nice. We now need only tables.c, tablec.h, util.c and util.h to
> compile the tables API with simplify.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/tskit-dev/msprime/pull/453#issuecomment-380770776>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AGHnHwSfLSVmVku9MgJLuGwjxlTajnhBks5tnzh7gaJpZM4TOIWB>
> .
>
-- 

Kevin Thornton

Associate Professor

Ecology and Evolutionary Biology

UC Irvine

http://www.molpopgen.org

http://github.com/ThorntonLab

http://github.com/molpopgen