@molpopgen 2018-04-12
It may be handy to have some separator comment lines in tests/__init__.py as it is getting really long.

For example, at the top, say:

```py
"""
You can search through this file by looking for three hash symbols (###), which I'm using here as breaks between sections
"""
```

Then, around line 620:

```py
### Python implementation of simplification algorithm
```