@molpopgen 2018-04-13
This is a first comparison of the new implementation in msprime vs what I've implemented [here](https://github.com/molpopgen/fwdpp_experimental).  The data size is approximately 81 million edges.

All timings are taken AFTER reading the data in from files and putting them into the relevant data structures.  So, my vectors of stuff are filled, and msprime's FooTables are populated, and then we sort and simplify.

No mutation simplification is done here.  (I've still not had time to return to that.)

Apologies for the different units for the time measurements--ms vs s.

@DL42 ran this, and will run more, on smaller data sets, pretty soon:

```
# Dump a large amount of data to binary files.  Binary avoids rounding errors that would prevent
# us from comparing results
./fwdpp_dump_nodes_edges 2000 1000 100 0.1 42 nodes.bin edges.bin

# Run a standalone C++ simplifier:
fwdpp_experimental dlawrie$ ./test_simplify nodes.bin edges.bin nodes_out.bin edges_out.bin 15
16763 ms
80004000 nodes, 80997299 edges
sort time = 9686 ms
simplify time = 2608 ms


python check_fwd_sim_data_with_msprime.py nodes.bin edges.bin 15 edges_out2.bin nodes_out2.bin
/Users/dlawrie/anaconda/envs/fwdpp_exp/lib/python3.6/site-packages/h5py/__init__.py:36: FutureWarning: Conversion of the second argument of issubdtype from `float` to `np.floating` is deprecated. In future, it will be treated as `np.float64 == np.dtype(float).type`.
  from ._conv import register_converters as _register_converters
/Users/dlawrie/anaconda/envs/fwdpp_exp/lib/python3.6/site-packages/msprime-0.5.1.dev32+g860a0a9.d20180412-py3.6-macosx-10.9-x86_64.egg/msprime/__init__.py 0.5.1.dev32+g860a0a9.d20180412
Sorting:  13.168102741241455 seconds
Simplifying:  8.899866104125977 seconds
```

I expect performance differences to shrink on smaller total data sizes, which I've seen before.